+++ index.css
@@ -8,11 +8,10 @@
 @import url(./src/widgets/interestsEdit/ui/index.css);
 @import url(./src/entities/person/ui/match/index.css);
 @import url("/src/pages/main/index.css");
-@import url("/src/widgets/navbar/ui/navbar.css");
-@import url("/src/widgets/header/ui/header.css");
+@import url("/src/widgets/navbar/ui/index.css");
+@import url("/src/widgets/header/ui/index.css");
 @import url("/src/entities/person/ui/card/index.css");
 @import url("/src/pages/landing/landing.css");
-@import url("/src/pages/matches/components/card.scss");
 @import url("src/pages/matches/matches.scss");
 @import url('/src/shared/ui/modal/index.css');
 @import url(/src/assets/fonts/MailSans/stylesheet.css);
