--- src/pages/landing/components/ticker/ticker.hbs
@@ -1,8 +1,0 @@
-<div class="ticker__container clearfix">
-    <div class="ticker__content clearfix">
-        бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом
-    </div>
-    <div class="ticker__content clearfix">
-        бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом бегом
-    </div>
-</div>
