+++ src/pages/layout/layout.tsx
@@ -10,8 +10,8 @@ import Header from '../../widgets/header/index';
 const Layout = ({ children }: any) => {
     return (
         <div className="layout">
-            {/* <Navbar /> */}
             <Navbar />
+            {/* {Navbar()} */}
             <div className="layout__content">
                 <Header />
                 <div className="layout__content__body">{children}</div>
