--- src/pages/matches/components/card.hbs
@@ -1,25 +1,0 @@
-<div class="match">
-            <button class="mail">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
-                    stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
-                    class="lucide lucide-mail">
-                    <rect width="20" height="16" x="2" y="4" rx="2" />
-                    <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-                </svg>
-            </button>
-            <div class="profile-pic profile-pic1">
-                <img src="{{image}}"
-                    alt="pic">
-            </div>
-            <div class="data">
-                <div class="data__content">
-                    <span class="name">{{name}}, {{age}}</span>
-                    <span class="about-me">{{description}}</span>
-                </div>
-                <div class="data-bottom">
-                    <button class="match__button">Contact Me</button>
-                </div>
-            </div>
-        </div>
-</div>
-
