--- src/pages/matches/components/card.scss
@@ -1,157 +1,0 @@
-.match {
-  $card7-color: darken($color: #fcd0ce,$amount: 5);
-  width: 300px;
-  height: 300px;
-  background: rgba(77, 192, 54, 0.404);
-  border-radius: 32px;
-  padding: 3px;
-  position: relative;
-  transition: all 0.5s ease-in-out;
-  box-shadow: #604b4a30 0px 70px 30px -50px, rgba(100, 100, 111, 0.103) 0px 0px 29px 0px;
-
-
-  .mail {
-    position: absolute;
-    right: 2rem;
-    top: 1.4rem;
-    background: transparent;
-    border: none;
-
-    svg {
-      stroke: $card7-color;
-      stroke-width: 3px;
-
-      &:hover {
-        stroke: darken($color: $card7-color, $amount: 20);
-      }
-    }
-  }
-  .profile-pic {
-    position: absolute;
-    width: calc(100% - 6px);
-    height: calc(100% - 6px);
-    top: 3px;
-    left: 3px;
-    border-radius: 29px;
-    z-index: 1;
-    border: 0px solid $card7-color;
-    overflow: hidden;
-    transition: all 0.5s ease-in-out 0.2s, z-index 0.5s ease-in-out 0.2s;
-
-    img {
-      object-fit: cover;
-      width: 100%;
-      height: 100%;
-      object-position: 0px 0px;
-      transition: all 0.5s ease-in-out 0s;
-    }
-  }
-
-  .data {
-    position: absolute;
-    bottom: 3px;
-    left: 3px;
-    right: 3px;
-    background: $card7-color;
-    top: 80%;
-    border-radius: 29px;
-    box-shadow: #604b4a30 0px 5px 5px 0px inset;
-    overflow: hidden;
-    transition: all 0.5s cubic-bezier(0.645, 0.045, 0.355, 1) 0s;
-
-    .data__content {
-      position: absolute;
-      bottom: 0;
-      left: 1.5rem;
-      right: 1.5rem;
-      height: 170px;
-      width: auto;
-
-      .name {
-        display: block;
-        font-size: 1.2rem;
-        color: white;
-        font-weight: 900;
-      }
-      .about-me {
-        display: block;
-        font-size: 0.9rem;
-        color: white;
-        margin-top: 1rem;
-        overflow-wrap: break-word;
-      }
-    }
-    .data-bottom {
-      position: absolute;
-      display: none!important;
-      bottom: 1rem;
-      left: 1.5rem;
-      right: 1.5rem;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      .social-links-container {
-        display: flex;
-        gap: 1rem;
-        svg {
-          height: 20px;
-          fill: white;
-          filter: drop-shadow(0 5px 5px #a5848222);
-
-          &:hover {
-            fill: darken($color: $card7-color, $amount: 20);
-            transform: scale(1.2);
-          }
-        }
-      }
-
-      .match__button {
-        background: white;
-        color: $card7-color;
-        border: none;
-        border-radius: 20px;
-        font-size: 0.6rem;
-        padding: 0.4rem 0.6rem;
-        box-shadow: #a5848222 0px 5px 5px 0px;
-
-        &:hover {
-          background: darken($color: $card7-color, $amount: 20);
-          color: white;
-        }
-      }
-    }
-  }
-
-  &:hover {
-    border-top-left-radius: 55px;
-    .data {
-      top: 20%;
-      border-radius: 80px 29px 29px 29px;
-      transition: all 0.5s cubic-bezier(0.645, 0.045, 0.355, 1) 0.2s;
-    }
-
-    .profile-pic {
-      width: 100px;
-      height: 100px;
-      aspect-ratio: 1;
-      top: 10px;
-      left: 10px;
-      border-radius: 50%;
-      z-index: 3;
-      border: 7px solid $card7-color;
-      box-shadow: #604b4a30 0px 5px 5px 0px;
-      transition: all 0.5s ease-in-out, z-index 0.5s ease-in-out 0.1s;
-
-      &:hover {
-        transform: scale(1.3);
-        border-radius: 0px;
-      }
-
-      img {
-        transform: scale(2.5);
-        object-position: 0px 25px;
-        transition: all 0.5s ease-in-out 0.5s;
-      }
-    }
-  }
-}
