--- src/pages/matches/match.hbs
@@ -1,9 +1,0 @@
-{{> components/header/header authState=1}}
-<div class="matches">
-    <h1 class="matches__header">Ваши мэтчи</h1>
-    <div id="matches__content">
-        {{#each matches}}
-            {{> pages/matches/components/card}}
-        {{/each}}
-    </div>
-</div>
