+++ src/shared/ui/input/input.tsx
@@ -1,7 +1,8 @@
 import { TSize, getClassBySize } from '../types';
-import { clsx } from '../../../clsx/index';
+import { clsx } from '../../lib/clsx/index';
 import { useState, useEffect } from '../../../reactor/index';
-import { genId, validateInput } from '../../lib/index';
+import { genId } from '../../lib/common/index';
+import { validateInput } from '../../lib/input/index';
 
 export type TInput = {
     autofocus?: boolean;
