+++ src/widgets/decriptionEdit/ui/index.tsx
@@ -2,7 +2,7 @@ import { Modal, Button, TextArea } from '../../../shared/ui';
 import { useEffect, useState } from '../../../reactor';
 import { updateDescription } from '../../../entities/session/api';
 import { store } from '../../../app/app';
-import { clsx } from '../../../clsx';
+import { clsx } from '../../../shared/lib/clsx';
 
 /**
  * A DescriptionEdit component that renders a form for editing the user's description.
