+++ src/widgets/footer/ui/index.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from '../../../reactor/index';
-import { Link } from '../../../shared/routing/link';
-import { clsx } from '../../../clsx/index';
+import { Link } from '../../../shared/ui/link';
+import { clsx } from '../../../shared/lib/clsx/index';
 import { store } from '../../../app/app';
 
 const Footer = () => {
@@ -58,7 +58,11 @@ const Footer = () => {
                     onClick={() => {
                         store.dispatch({
                             type: 'UPDATE_CURRENT_CHAT',
-                            payload: null,
+                            payload: {
+                                id: 0,
+                                name: "",
+                                photo: "",
+                            }
                         });
                     }}
                     style="font-size: x-large"
