+++ src/widgets/navbar/ui/index.css
