+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE = 'jimder-cache-v11';
+const CACHE = 'jimder-cache-v26';
 const PRECACHE_URLS = ['/offline', '/176c4714b229b0ae6633.webp'];
 
 const NETWORK_ONLY = ['isAuth', 'cards', 'getAllChats'];
